backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "aiomysql>=0.3.2",
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.135.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-milvus>=0.3.3",
    "llama-index>=0.14.20",
    "mysql-connector-python>=9.6.0",
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.5.0",
    "pydantic-settings>=2.13.1",
    "pymilvus>=2.6.12",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "pandas>=3.0.2",
    "openpyxl>=3.1.5",
    "sentence-transformers>=5.4.0",
    "langchain-text-splitters>=1.1.1",
    "langchain-openai>=0.3.7",
    "prometheus-fastapi-instrumentator>=7.0.0",
]

[tool.uv]
# Force uv to look at the PyTorch CPU index FIRST to prevent downloading the 5GB NVIDIA CUDA GPU wheels
index-strategy = "unsafe-first-match"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
